CPR-79 Stop sending null in string to match score
