+++ src/test/kotlin/uk/gov/justice/digital/hmpps/personrecord/message/listeners/cpr/ReclusterEventListenerIntTest.kt
@@ -3,7 +3,7 @@ package uk.gov.justice.digital.hmpps.personrecord.message.listeners.cpr
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
-import uk.gov.justice.digital.hmpps.personrecord.client.MatchResponse
+import uk.gov.justice.digital.hmpps.personrecord.client.model.match.MatchResponse
 import uk.gov.justice.digital.hmpps.personrecord.client.model.offender.Identifiers
 import uk.gov.justice.digital.hmpps.personrecord.client.model.offender.Name
 import uk.gov.justice.digital.hmpps.personrecord.client.model.offender.ProbationCase
